requirements.txt
lxml
python-slugify
Flask